Restore from ignore
